Fix bug causing code to download incorrectly
The bug:
Open a sketch in view-only mode. Go to the sketches page and try to download your most recently viewed sketch. The sketch title and extension are as expected, but the code is the code from the view-only sketch.

This fixes the bug by saving the original code for the most recently viewed program, and restoring it before the component unmounts.