Refactored Output.js component to avoid using an <iframe> when running Python code (#759)
Made the Skulpt library draw directly to the canvas when running Python turtle code and write directly to the output container element when printing to the console. The code still uses <iframe> for the others (namely React and Processing.js), but I haven't been able to find a suitable online editor for those from which to draw inspiration.